+++ include_private/ASCAN/390N_T8/ldat.hpp
@@ -214,7 +214,7 @@ namespace Union::__390N_T8 {
         virtual double                      getBaseGain(int idx) const override;
         virtual double                      getScanGain(int idx) const override;
         virtual double                      getSurfaceCompentationGain(int idx) const override;
-        virtual int                         getSupression(int idx) const override;
+        virtual double                      getSupression(int idx) const override;
         virtual Union::AScan::DistanceMode  getDistanceMode(int idx) const override;
         virtual std::optional<Base::AVG>    getAVG(int idx) const override;
         virtual std::optional<Base::DAC>    getDAC(int idx) const override;
